docker/mcp/requirements.txt
mcp>=1.0.0
requests>=2.31.0
qdrant-client>=1.7.0
pyyaml>=6.0
